Started used new output functionality around
In particular, the command line options parser has been refactored
extracting some functions in an auxiliary file. Some work was done on
the helper to make it not too wide and consistent regarding colors.
It was decided to delete all the Juqueen related code; I started doing
so but some code was temporarily left commented.