Worked on checks of variables before execution
The user configuration file of BaHaMAS contains quite a bit of variables
and it could be annoying to be obliged to fill it out completely,
especially if not all the functionality of BaHaMAS are going to be used.
Still, variables must be set and they should never be used if unset
(this could be kind of automatically done by 'set -u' but it is not
user-friendly, especially if the user is not a Bash expert).
Therefore, I decided to make checks on variables before starting the
execution of the code (i.e. after having parsed the command line
options), checking different variables depending on the use of BaHaMAS
chosen by the user. This will be less easy to maintain, but it helps the
user, giving readable error messages. The code must still be test and it
should be done with 'set -euo pipefail'.