Refactored main code improving structure
The database case was treated as the others and the case in which no
mutually exclusive option is given to BaHaMAS has now become an 'else'
case of the long if-elif-else-fi chain. This allowed to keep the
variable storing the mutually exclusive option passed local in the
command line option parser.