Generalised command line options parsing
The idea is that the user is completely free to specify the command line
options with the preferred style. A space or an equal sign can be used
to separate an option from its value and even a mix of those works. It
is then understood that a first equal is not considered as part of an
option and, if this is required, a double equal must be specified to
specify the option (this is a remote case). The parser was also
generalised to include option values with spaces, even though at
the moment it is not needed.