Added checks on system requirements
Since there is not really an universal way of getting veersions from
programs, at the moment we checked only bash, awk and sed versions and
for the last two the check could be skipped on some clusters. Maybe we
will come up with a better way to do these checks.