Activated bash strict mode and fixed failing tests
Using in particulare the set -eu options, it turned out that many
undefined/unset variable were used. Often this was ok, sometimes it was
wrong. We want to continue in this setup and we do not exclude that
there are bugs around in the code that were not caught by the tests,
just because the existing tests do not cover 100% of the code.