Allowed less strict mode in loading user variables
The check on Wilson/Staggered variables has been moved again after
helpers printing, so that the user can ask the helpers from everywhere.
Doing so, it could be that variables are unset and therefore in helpers
they would not be printed. We made the helper print a 'unset' keyword in
orange to let the user know.