Changed betas files conventions and syntax
Now in the betas file, except from the betas value that must appear in
first position on each line without prefix, all the other entries need
to be prefixed (e.g. 's' for seed). The order of these entries in the
line is completely arbitrary. Still some work has to be adapted to the
--completeBetasFile option which now does not do what it should (even if
the tests pass). New entries in betas such as time per trajectory and
goal in statistics have been introduced and will be soon considerend in
the submitting part of BaHaMAS.