Incorporated jobs monitoring into BaHaMAS with workaround
The option --jobstatus is not treated as the others mutually exclusive
options at the moment, since we would like this functionality to work,
no matter from which position the script is called (afterall it is
nothing more than a parser of information of the scheduler). But BaHaMAS
makes checks on many aspects which should not be done (or done
differently) in the jobstatus scenario.