Continued working on exit codes in case of errors
Whenever something was printed to screen before exiting, a proper exit
code as well as the Fatal function has been used. This required some
adjustments of the code in order to keep the output as it was. The
database parser has been improved extracting some printing functionality
which is in common with all the parsers.