Look for last valid checkpoint in a more solid way
Before, when the user was using 'rlast' in the betas file, the last
configuration was looked for and, if the prng file was not found, then a
random seed was used. Now, if the user asks 'rlast', it is better to
think of this as last valid checkpoint, meant as PAIR of files!