Fixed bug due to unrequired variable to be filled
In the simulation status code for SLURM the standard output of the
CL2QCD code is used to get the time per trajectory. But this operation
is done only if the user explicitly asks it and in particular the
database does not need to use the output file. So the variables related
to the secondary option had to be declared locally inside the if case.