Add options to modify nflavor and seed prefixes
This was for completeness. Indeed, all prefixes were declared as
readonly at declaration time and the existing options to possibly
change them were useless and they were resulting in errors. Now, the
prefixes variables are marked as readonly in the command line parser
which needs to modify also prefixes-related variables in case the
prefix is given by the user.