Fix serious bug in producing input file for SLURM
The value of the chemical potential on the Roberge-Weiss plane in
lattice units is mu=pi/Nt and, thus, changing Nt affects the value of
mu. Nevertheless, mu was fixed to pi/6 and it was not depending on Nt.
This implies that previous simulations submitted with BaHaMAS at
mu=muPiT and not at Nt=6 were not run on the RW plane.