Fix subtle bug in SLURM simulation status
In the commit 7fc993033e55f3393087c038ec4fec3870c8a0cd some work on
error codes was done, in order to use global constant variables to
encapsulate the error codes. However, in the awk script to check the
correctness of the CL2QCD output, the exit codes were not initialised
and, therefore, 0 (success) was always returned.