Fix bug in setup
If a new variable was introduced in the user variables template, the
setup was failing with an unbound variable error. Fixing it, a new
interactive "press enter" step was added in order to make the user be
aware of missing variables. This situation might be normal and going on
in the setup should let BaHaMAS add the missing variables.