Improve jobscript preparation
Some old naming variables were renamed to small characters and were
made local. The walltime calculation was improved also with respect
to error handling. The printing of the grouped betas in the same job
was delegated to the jobscript preparation function together with the
printing of the exclude node string.