Introduce string execution mode variable
So far variables either equal to TRUE or to FALSE were used to depict
execution modes and this grew up over time. It makes sense to call the
modes in a fixed way and use just one variable. The string to call the
modes was chosen to have 'mode:' as prefix.