Move and improve some independent code to own folder
Some functions were marked as SLURM specific, but this was not the
case. Some were containing particular calls to CL2QCD specific code,
but this dependence has been made general using the software variable
in the functions call. The resulting code was generic and thus it has
been moved to the appropriate folder.