Distribute simulation status code to own folders
Most of the simulation status code was not really dependent on the
scheduler and the only part which was interacting with it was that
extracting the status of the queued jobs, used in turn to determine
whether the considered beta folders were referring to queued
simulations. For the rest, lots of code was CL2QCD specific and
implicitly based on the format of the input and output files. This has
been delegated to a function which will have to be implemented for each
LQCD software.