Make SLURM function name and implementation consistent
The functions that gather jobs information were not all called with a
similar name logic and their implementation was not consistent. Now the
job information retrieved from the scheduler is put in the
jobsInformation array and each of its entry is a @-separated list of
information. This is parsed later depending on the following needs.