Breaking change refactoring of UI of BaHaMAS
The command line options have been radically refactored. Mutually
exclusive options have been replaced by execution modes. BaHaMAS
accepts now only in first position the mode in which it should be
executed. The --help option can be used to get help either in general
or about the previously specified mode. The helper have not yet been
adjusted as well as lots of old code in the MainParser.bash file has
still to be removed. This are going to be next steps.