Implement specific mode parsers and option checks
The parsing of the command line options has been deeply refactored. A
parser per mode can be now easily modified or added. Options which are
common to several modes are accepted only if really used and not
ignored as before. The help is now outdated and it will be worked on in
the next changes. The code in the MainParser.bash file has been on
purpose put in a disordered way, in order to avoid that git mixes new
and old code in the diff. In next commit we will adjust it.