Refactor parsers such that no mode can accept a value
New options for some modes have been introduceed to avoid that values
can be passed to modes. This makes a further distiction between modes
and options and it allows for better handling of equals on the command
line. In particular, now it is possible to strictly respect the rule
that an optional equal sign can be used to specify the value of an
option with arbitrary space around him. Such an equal sign will be
ignored at the very beginning of the execution.