Add initial support for manuals with automatic creation
The BaHaMAS manual has been created and the infrastructure for manuals
has been created. Giving '--help' after a mode will create manuals from
the available markdown files and then try to invoke the asked one, if
available. Still the BaHaMAS manual is not automatically open by the
codebase at any stage and it will be then decided in which way to go.