Move function to give help to its file and improve it
In particular, a function to create manuals has been extracted and
called both when any help is asked and when the setup is run. Being the
creation of manual pages done using a Makefile, there should not be any
overhead, because make recreates only pages whose source has changed
and this should not be the case for the end user.