Avoid dependence on pandoc and distribute manuals
Making manuals at "help time" required a dependence on pandoc, which
add complexity for the user basically for nothing. It is better to
leave to the developer to ensure manuals are produced and shipped
within the repository. To be on the safe side, a check for
not-up-to-date manuals has been implemented and a hidden mode (i.e. not
documented) has been added to update manuals. At least, in case the
developer forgets to make the manuals, the users can be warned and even
fix it, if they can (i.e. have pandoc installed).