Extracted scheduler and software code and fix bug
The production job script was entirely in the CL2QCD folder, but it was
containing also scheduler specific directives. The code has been then
refactored. A bug was also fixed: Previously we introduced braces in
all parameter expansions, but in the job script this was done too
naively, because there the expansion is escaped in BaHaMAS to do it at
a later point in the job script and the variable name is then
constructed using an index in it. The closing braces was in the wrong
place, e.g. "\${dir}${index}" instead of "\${dir${index}}".