Deeply refactor tests to increase code coverage
Now tests can be easily cover the case where in the mocked environment
the submit disk is different than the run one. This was not the case
before and therefore now the functional tests consider a less trivial
case. Also different projects folder and different parameters have been
used for CL2QCD and OpenQCD, in order to emulate both staggered and
Wilson fermions in real life.