Do not delegate executable/input files  copy to job script
So far, for CL2QCD, the job script was responsible to copy to the run
folder the executable and the input files. This is not really needed
and it will in any case not be possible with OpenQCD where the code has
to be compiled for every job and it has to handled by BaHaMAS. New code
was added to produce the executable file (at the generic level calling
a software specific function through the interface) and to ensure that
on the run disk the needed files are present. This last step is clearly
not needed if the run disk is the same of the submit disk. Finally,
having changed this aspect, the logic in the submit-only mode has been
adapted (more checks should then be done) and the copy in some
modes (continue* and measure) is limited either to the input or to the
executable file. Tests have been adjusted accordingly.