Generalize SLURM jobs directives to consider MPI jobs
To keep the SLURM code independent from software-specific code, the
distiction between GPU and MPI job is done checking if the variable to
specify the number of GPU per nodes is set. This is not very flexible
and it is quite ad hoc for the moment. A refactoring will be needed in
future to make the code more general.