Change approach in creation of production executable file
Compilation is not required for every beta and hence it is now done
only once. The function that creates the executable does it for all
beta folders. For codes to be compiled, the processor grid is now part
of the executable name in order to let the user not have to specify it
for e.g. the continue mode and in general to let BaHaMAS to retrieve
it in case of need.