Extract processor grid from executable name in continue mode
Although this complicates a bit the logic of the codebase, it makes
sense to allow the user not to have to always give the processor grid
on the command line. Since openQCD should not be used to continue jobs
changing the processor grid, this is put in the executable name at
submission and read off at continuation.