Read processor grid from executable in submit-only, too
In order to correctly check the needed files for the submission, the
executable file name is needed but in submit-only mode the processor
grid must not be specified by the user and hence it is now read from
the executable name, which is clearly supposed to exist. We also
decided to make the global variable containing the executable name
readonly after that it is set in all modes, hence in the validation of
the beta values after having parsed the betas file.