Add run-time renaming mechanism to openQCD job script
The openQCD software does not use the same convention of CL2QCD about
how to store checkpoints and in particular the random number generator
is overwritten at every checkpoint, what in finite T LQCD implies that
a run cannot be properly resumed from a previous trajectory. A quite
complex mechanism has been devised to bring openQCD checkpoints to the
same form of CL2QCD ones. This is going to facilitate the continue mode
implementation.