Complete implementation of thermalize mode and add tests
Basically the missing brick was the backup of the thermalized
configuration to the dedicated folder and this was added to the
jobscript. For openQCD (and in future it will be changed also for
CL2QCD) the name of the thermalized configuration has been improved in
order to contain the number of the trajectory to which the
configuration refers. Although this let us loose the information of how
many trajectories were done from hot in case of thermalized
configurations from conf, it will allow to simplify a lot the meaning
of the g field in the betas file which at the moment changes from new
run to continuation.