Fix checkpoint rename mechanism to consider start trajectory
The trajectory number to rename checkpoints for openQCD was determined
with the gap between trajectories only, but not using the initial
trajectory number, too. Hence, a thermalization fromConf or a new chain
were having chekpoints starting from zero again. This has been fixed,
although it was not trivial to decide were to take the initial
trajectory number. For the moment this is taken from the initial
configuration filename which however requires a branch between
execution modes. If the simulation is being started, then the
configuration from the pool is set and used; if the simulation is being
continued, then the symbolic link to the initial configuration to the
pool is used. Note that the symbolic link is not there for new runs at
job script creation (it is created later) and it is then not always
possible to use the symbolic link.