Fix behaviour of CL2QCD continue to make g field consistent
The way the number of done trajectories is deduced has been
changed. Before it was referring to the number of trajectories done in
total from hot and this was forcing the user to consider in the g field
in the betas file the number of trajectories done in the thermalization
as well. Now this is not needed and the user must simply specify how
many trajectories in the new chain should be done.