Store openQCD log file on run disk and backup at job end
The log file in openQCD is more a data file than a standard
output (this was the reason it was stored at first on the submit
disk). However it has to be cleaned in continue mode in case of a
resume and it makes sense to treat it as the output file in CL2QCD
which is then stored on the run disk and copied back to the submit disk
at the end of the simulation.