Add mechanism to restore environment in case of later error
The continue mechanism modifies files quite a lot and it would be nice
to avoid triggering later artificial failures of BaHaMAS when it has to
be re-run due to a given error. In openQCD this was not the case, since
the files of the last checkpoint were renamed and were remaining so if
a failure occurred after the rename operations. Running BaHaMAS again
would have fail because the checkpoint would have been seen as failure
in the renaming mechanism in the previous job. Now this should be
fixed, since we rename back the checkpoint files to the standard
BaHaMAS naming scheme in case of any failure in continue mode. For
CL2QCD nothing should be done and the restore function is a no-op.