Refactor and improve continue mode for all software
The separation of resposibility has been made stronger and there is now
one function looking for the checkpoint to be used, that does not
modify the state of the beta folder, and two functions which act either
on output files or on the input file. This resposibility spearation
allowed to easily implement a mechanism to restore the state of the
betas folder in case of later failuers, e.g. moving files back to the
beta folder from the trash folder. In this way the user experience
should be way better. Few adjustments to tests and comments have been
made. The trash folder path has been moved to the caller function to
allow the restore mechanism to work.