Make simulation status work on universal output file
It has been decided to let the simulation-status mode create a generic
format output file, which is then not relying on any software. CL2QCD
format has been chosen to be also the standard software. This universal
output file will have the same name as the simulation output file but
with a .BaHaMAS extension. For CL2QCD a symbolic link to its output
file is created, because the format is by definition the needed one.