Add command line option to change number of cores per node
The user might choose a processors grid for openQCD which does not make
use of full nodes and still have the total number of cores per node in
the setup. This would lead to inconsistent and potentially wrong
directives in the job script and the submission might be rejected by
the scheduler. Having a new command line option allows the user to take
control. Indeed, now, if the user specifies a processors grid which
would leave some cores unused, an error is printed and the number of
cores per node is reduced, till it becomes a divisor of the total
number of cores to be used.