Check parameter regex in path at extraction time
The check if a parameter extracted from path was matching the expected
regex was done after having extracted its value from path. In this
operation the string between the prefix and the following slash was
taken and it was possible that the user had by accident a folder name
starting with the prefix as well. An unwanted error could occur.
Now the parameter is extracted matching directly prefix and regex at
once. No check is then needed afterwards.