Fix bug about running BaHaMAS before the firt setup
The user should be able to run BaHaMAS to get help or the version
without having to do the setup. This last case was not considered and
now it has been fixed. Running BaHaMAS without command line options
should be equivalent to run it with --help and this has been moved at
the very beginning to avoid further logic at later points.