Add new setup variable to choose command to run software
It might happen that the user needs to use something different than
srun on a cluster and this is indeed something that should be decided
by the scheduler, if not given by the user, and not be hard-coded in
the job script file. A new setup variable has been introduced and a
default command (e.g. srun for SLURM) will be used if not specified by
the user in the setup.