Fix bug in calculation of time elapsed from last trajectory
In simulation-status mode the time from the last trajectory was wrong
because it was calculated using the last modification time of the
freshly created standardized output file and not using that of the
software real output file. The problem has been fixed changing the last
modification time of the standardized file to that of the software
output file.