Add smart sleeping time calculation for openQCD job script
Before the initial sleeping time was hard-coded to 10s and this in
continue mode could have led to rename the resume checkpoint before
that openQCD read it with a consequent error in the run (configuration
was then not found). Now, in continue mode, the time between
checkpoints is estimated using information from the input file and the
log file which must be present in any case. Then the initial sleeping
time is set to half of the time between the checkpoints.