Quote expansions calling functions to act on input file
In the continue mode operations in the software implementations the
function to edit the input file were called without quoting the
expansions handed over and, although they were not containing globbing
characters, a bug of the version 5.0.0(1) of bash was hit if in the
string a backslash was occurring. This was activating filename
expansion and the nullglob option of the shell was the handing over an
empty string to the replacing function, which was not changing anything
in the input file. Quoting the expansions forbids globbing from
occurring at any stage.