Improve system requirement checks and add report to help
The code to check the system requirements has been refactored in order
to be able to add new requirements in the feature simply by adding an
entry to a global associative array. The general BaHaMAS help gives now
a report to the user about the versions of the needed software.