Improve CL2QCD functional tests to consider two simulations
In real life more than one simulation is done in a single job when
running CL2QCD on GPUs. Functional tests are now dealing with this more
real scenario and, indeed, a bug due to a wrong usage of the global
variable BHMAS_betaValuesToBeSubmitted is now reproducible and a bunch
of tests fail.