Fix bug due to usage of wrong global variable
In some functions in Generic_Code/ProduceFilesForEachBeta.bash the
wrong run IDs were used to build up beta folders and this was
triggering errors only when using more than one simulation per job,
because there the string to build up beta folder names was containing
more than one seed, hence producing not existing folder names. Now a
global variable with a better name is created and used. Some names
around are still not ideal, especially because often a beta value is
indeed a run ID (beta, seed, postfix) but it is already planned to
improve this aspect in future.