Add new option to update executable file in continue modes
By default, the continue and the continue-thermalization modes are
requiring that the executable file is present in each beta folder, but
it might as well be that the user needs to update it (e.g. because of a
software update by the administrators). This is now possible and few
functional tests have been added.