Fix --local option of job-status execution mode
The ooption was broken due to an unquoted expression which was then
washed away by glob expansion together with nullglob being set. The fix
has been done actually moving the implementation in the general code
and not in the SLURM code, since this can be done and make the code
easier in future to be expanded. Some functional test has been added,
mocking the squeue function in a better way, depending on the test
being run.