Fix bug about ignored setup variable in continue* modes
The BHMAS_useRationalApproxFiles set to FALSE was ignored in the
continue modes and the input file lines with the specification of the
rational approximation files where attempted to be modified, but
potentially and probably not there. It is now also considered
the (remote) case where the user starts simulations with
BHMAS_useRationalApproxFiles set to FALSE, i.e. without having the
rational approximation files paths in the input file, and then at a
later point changes the setup variable to TRUE. In this case the lines
would be added accordingly.