Add pseudofermions field to betas file
This has been implemented in an analogous way to the g field and it
required a new associative global array to keep track of the values
specified in the betas file. The input file preparation and the
continue* execution modes have been adjusted and this new feature has
been included in the tests betas file.