Make acceptance-rate-report mode act on existing folder
Before this mode was constrained to act on new-chain only, but this is
not ideal and it was now adjusted to be able to act on a thermalization
if no new chain folder exists. This must be done as early as possible
since the ID label is then used in keys of associative arrays. Hence it
has been done in the beta file parsing as soon as the beta value is
read. This is done only for the acceptance-rate-report mode, though.