Fix bug about erroneous deleting of executable with CL2QCD
Since openQCD support was added, executable files are copied to the
beta folder either from the compilation folder (openQCD) or from the
reference place (CL2QCD). In continue* mode this is done only if the
user uses the --updateExecutable option. However, the CL2QCD production
job was deleting the executable and this was making following continue
jobs failing.